requirements.txt
googlesearch-python
duckduckgo-search
streamlit